move isAmbient to transformer_util
This reduces some circularity in our imports, and also is a more
logical place for it.
